modified read_size type in fopen

+++ get_next_line/get_next_line_bonus.h
@@ -27,7 +27,7 @@
 
 typedef struct fopen
 {
-	int		read_size;
+	ssize_t	read_size;
 	char	*buff;
 	size_t	str_size;
 	int		is_newline;
